+++ src/components/results/search-hit/SearchHit.tsx
@@ -1,32 +1,13 @@
 import React from 'react';
 import htmlReactParser from 'html-react-parser';
-import { formatDate } from 'utils/datetime';
-import dayjs from 'dayjs';
 import { Audience, SearchHitProps } from 'types/search-result';
 import { logResultClick } from 'utils/amplitude';
-import { BodyLong, BodyShort, LinkPanel } from '@navikt/ds-react';
+import { BodyLong, LinkPanel } from '@navikt/ds-react';
 import { SearchHitOfficeInformation } from './office-information/SearchHitOfficeInformation';
 import { SearchHitAudience } from './audience/SearchHitAudience';
+import { SearchHitBottomText } from './bottom-text/SearchHitBottomText';
 
 import style from './SearchHit.module.scss';
-
-const createPublishedAndModifiedString = ({
-    publish,
-    modifiedTime,
-    createdTime,
-}: SearchHitProps) => {
-    const publishedTime = publish?.first || createdTime;
-    const publisedString = publishedTime
-        ? `Publisert ${formatDate(publishedTime)}`
-        : '';
-    const modifiedString =
-        modifiedTime && dayjs(modifiedTime).unix() > dayjs(publishedTime).unix()
-            ? `${publisedString ? ' | ' : ''}Sist endret ${formatDate(
-                  modifiedTime
-              )}`
-            : '';
-    return `${publisedString}${modifiedString}`;
-};
 
 const parseHighlight = (highlight: string) => {
     return htmlReactParser(
@@ -46,7 +27,6 @@ export const SearchHit = ({ hit, hitIndex, searchTerm }: Props) => {
         displayName,
         href,
         highlight,
-        priority,
         officeInformation,
         audience = (['person', 'employer', 'provider'] as Audience[])[
             Math.floor(Math.random() * 3)
@@ -56,8 +36,6 @@ export const SearchHit = ({ hit, hitIndex, searchTerm }: Props) => {
     if (!displayName || !href) {
         return null;
     }
-
-    const publishedString = createPublishedAndModifiedString(hit);
 
     return (
         <LinkPanel
@@ -76,24 +54,7 @@ export const SearchHit = ({ hit, hitIndex, searchTerm }: Props) => {
                     <SearchHitOfficeInformation {...officeInformation} />
                 )}
                 <div className={style.bottomRow}>
-                    <div className={style.bottomLeft}>
-                        {publishedString && (
-                            <BodyShort
-                                className={style.published}
-                                size={'small'}
-                            >
-                                {publishedString}
-                            </BodyShort>
-                        )}
-                        {priority && (
-                            <BodyShort
-                                className={style.recommended}
-                                size={'small'}
-                            >
-                                {'Anbefalt innhold'}
-                            </BodyShort>
-                        )}
-                    </div>
+                    <SearchHitBottomText hit={hit} />
                     {audience && <SearchHitAudience audience={audience} />}
                 </div>
             </div>
